move input and grid component to singlelineScoreInput file
